crud finalizado

+++ src/script.js
@@ -16,6 +16,8 @@ const btnAdd = document.querySelector(".btn-add");
 
 const tableUsers = document.querySelector(".table-users");
 
+let id;
+
 const renderUser = (doc) => {
   const tr = `
       <tr data-id=${doc.id}>
@@ -37,6 +39,13 @@ const renderUser = (doc) => {
 
   btnEdit.addEventListener("click", () => {
     editModal.classList.add("modal-show");
+
+    id = doc.id;
+
+    editModalForm.firstName.value = doc.data().firstName;
+    editModalForm.lastName.value = doc.data().lastName;
+    editModalForm.phone.value = doc.data().phone;
+    editModalForm.email.value = doc.data().email;
   });
 
   //Delete user
@@ -44,23 +53,36 @@ const renderUser = (doc) => {
     `[data-id= '${doc.id}'] .btn-delete`
   );
   btnDelete.addEventListener("click", () => {
-    db.collection("users")
-      .doc(`${doc.id}`)
-      .delete()
-      .then(() => {
-        console.log("Document successfully");
-      })
-      .catch((error) => console.log("Error removing document ", error));
+    const confirm = window.confirm("Do you really want to remove the selected user?");
+
+    if (confirm) {
+      db.collection("users")
+        .doc(`${doc.id}`)
+        .delete()
+        .then(() => {
+          // console.log("Document successfully");
+        })
+        .catch((error) => console.log("Error removing document ", error));
+    }
   });
 };
 
 btnAdd.addEventListener("click", () => {
   addModal.classList.add("modal-show");
+
+  addModalForm.firstName.value = "";
+  addModalForm.lastName.value = "";
+  addModalForm.phone.value = "";
+  addModalForm.email.value = "";
 });
 
 window.addEventListener("click", (e) => {
   if (e.target === addModal) {
     addModal.classList.remove("modal-show");
+  }
+
+  if (e.target === editModal) {
+    editModal.classList.remove("modal-show");
   }
 });
 
@@ -78,13 +100,36 @@ firebase.initializeApp(firebaseConfig);
 
 const db = firebase.firestore();
 
-db.collection("users")
-  .get()
-  .then((querySnapshot) => {
-    querySnapshot.forEach((doc) => {
-      renderUser(doc);
-    });
+// //Get all users
+// db.collection("users")
+//   .get()
+//   .then((querySnapshot) => {
+//     querySnapshot.forEach((doc) => {
+//       renderUser(doc);
+//     });
+//   });
+
+//Real time listener
+db.collection("users").onSnapshot((snapshot) => {
+  snapshot.docChanges().forEach((change) => {
+    if (change.type === "added") {
+      renderUser(change.doc);
+    }
+
+    if (change.type === "removed") {
+      let tr = document.querySelector(`[data-id='${change.doc.id}']`);
+      let tbody = tr.parentElement;
+      tableUsers.removeChild(tbody);
+    }
+
+    if (change.type === "modified") {
+      let tr = document.querySelector(`[data-id='${change.doc.id}']`);
+      let tbody = tr.parentElement;
+      tableUsers.removeChild(tbody);
+      renderUser(change.doc);
+    }
   });
+});
 
 addModalForm.addEventListener("submit", (e) => {
   e.preventDefault();
@@ -95,4 +140,15 @@ addModalForm.addEventListener("submit", (e) => {
     email: addModalForm.email.value,
   });
   modalWrapper.classList.remove("modal-show");
+});
+
+editModalForm.addEventListener("submit", (e) => {
+  e.preventDefault();
+  db.collection("users").doc(id).update({
+    firstName: editModalForm.firstName.value,
+    lastName: editModalForm.lastName.value,
+    phone: editModalForm.phone.value,
+    email: editModalForm.email.value,
+  });
+  editModal.classList.remove("modal-show");
 });
